bosskeys + multitoon updates
Bosskeys there was an attempt to make the goofy speedway racecar keep moving forward when window was not active.  Did not work

Multitoon bumped up the beans buying count to 250, and also tried tuning the launch function again.  Now it will try to double click whenveer clicking on ttr launcher just in case it wasn't the active window